SVG logo fill gradient fix
Last commit didn't fix it, this one should though. Had to move
url(#gradientID) inline instead of in CSS. The css fill url is relative
to the css file. Can be fixed with leading slashes but that only fixes
the root url. Best solution is inline since colors won't be changing